+++ {{ cookiecutter.project_name }}/composition_repository/django_builtins/settings.py
@@ -93,6 +93,7 @@ class DjangoBase(EnabledApplicationMarker):
         "django.contrib.auth.middleware.AuthenticationMiddleware",
         "django.contrib.messages.middleware.MessageMiddleware",
         "django.middleware.clickjacking.XFrameOptionsMiddleware",
+        "django.contrib.sites.middleware.CurrentSiteMiddleware",
     ]
 
     ROOT_URLCONF = "project.urls"