+++ {{ cookiecutter.project_name }}/composition_repository/project_sitemaps/settings.py
@@ -1,0 +1,15 @@
+from project_composer.marker import EnabledApplicationMarker
+
+
+class SitemapsSettings(EnabledApplicationMarker):
+    """
+    Sitemaps settings
+    """
+    @classmethod
+    def setup(cls):
+        super(SitemapsSettings, cls).setup()
+
+        cls.INSTALLED_APPS.extend([
+            "django.contrib.sitemaps",
+            "project_sitemaps",
+        ])